[ALPS09012070] input: touchscreen: kernel 6.6 : Fix build warning
[Description]
Fix build warning to error

[Test]
test ok and build pass

MTK-Commit-Id: 1e7e935086e0c1116536292721c028bd1dd48889

CR-Id: ALPS09012070
Feature: touch_driver

+++ drivers/input/touchscreen/gt9xx/gt9xx_update.c
@@ -498,7 +498,7 @@ cfg_fw_err:
 static u8 gup_check_firmware_name(struct i2c_client *client,
 		u8 **path_p)
 {
-	u8 len;
+	u16 len;
 	u8 *fname;
 
 	if (!(*path_p)) {
